Sample hook plugin to associate Git commits to Jira.
Allows to extend the existing Gerrit mechanism
to link Git commits to issue-tracker (commentLink)
by recognising the Jira issue and adding a comment
to back to Git commit.

In order to use this plug-in simply configure
commentLink to Jira and additional Jira connectivity
details into your gerrit.config file.
(see the plug-in documentation for more details)

Change-Id: I10328983377ac9cd8c85668098c712feeafa31c5