[RFC] Jira hooks redesigned on Gerrit 2.6 commit validation.
Completely new redesigned version of Jira hooks with
a common reusable infrastructure to build the issue-tracking
association logic to other Issue-Trackers (i.e. BugZilla).

The generic issue-tracker association maven module
is hooks-its and provides the logic for:
> Insertion of comment links using regex patterns
> Enforcement of issue-ids in git commits using
  Gerrit 2.6 commit validation listeners
> Automation of issue-tracker status transition 
  based on Gerrit code-review status

hooks-its does not make assumption on the underlying
issue-tracker system but provides only an abstract
facade interface ItsFacade to access and perform
actions on the issue-tracker.

The hooks-jira layer provides the implementation of
the ItsFacade and the init-step to configure and
check connectivity with Atlassian Jira.

Change-Id: I897bc98648ed24c723514f1701eb42c22acbb198