Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

Since this information is in the documentation of hooks-its, would it not be better to refer to that?  If it's manually added here it is easy to get out of date.

It would be good if there was a way to make the build of a plugin derived from hooks-its automatically include the hooks-its documentation.

Patch-set: 1
Label: Code-Review=-1